+++ packages/react-components/src/useWalletConnectorSelector.ts
@@ -30,14 +30,10 @@ export interface WalletConnectorSelector {
 }
 /**
  * React hook for managing a connector selector (usually a button in the UI).
- * To maximize flexibility, the basic {@link WithWalletConnector} tries to add as few constraints as possible
- * on how the connectors are used.
- * To implement a common case, this hook adds the additional rule that only one connector can exist at any given time.
- * Connections not belonging to this connector are automatically disconnected.
  *
  * More precisely, the hook computes the derived state {@link WalletConnectorSelector} from {@link props} as follows:
- * - The selector is {@link WalletConnectorSelector.isSelected selected} if the active connector has type {@link connectorType}.
- * - The selector is {@link WalletConnectorSelector.isConnected connected} if it is selected and has an active connection.
+ * - The selector is {@link WalletConnectorSelector.isSelected selected} if the active connector's type is {@link connectorType}.
+ * - The selector is {@link WalletConnectorSelector.isConnected connected} if it is selected and the active connection was created by the active connector.
  * - The selector is {@link WalletConnectorSelector.isDisabled disabled} if it there is another connected selector.
  *
  * It also exposes a {@link WalletConnectorSelector.select handler function} for connecting/disconnecting appropriately
@@ -51,14 +47,13 @@ export function useWalletConnectorSelector(
     connectorType: ConnectorType,
     props: WalletConnectionProps
 ): WalletConnectorSelector {
-    const { disconnectActive, activeConnectorType, activeConnector, activeConnection, setActiveConnectorType } = props;
+    const { activeConnectorType, activeConnector, activeConnection, setActiveConnectorType } = props;
     const isSelected = activeConnectorType === connectorType;
-    const select = useCallback(() => {
-        disconnectActive();
-        // TODO Use isSelected here and remove redundant values from the list of dependencies.
-        setActiveConnectorType(activeConnectorType === connectorType ? undefined : connectorType);
-    }, [connectorType, activeConnector, activeConnectorType]);
-    const isConnected = Boolean(isSelected && activeConnection); // the hook's semantics ensures that activeConnection originates from this connector
-    const isDisabled = Boolean(activeConnectorType && activeConnectorType !== connectorType && activeConnection);
+    const select = useCallback(
+        () => setActiveConnectorType(isSelected ? undefined : connectorType),
+        [isSelected, connectorType]
+    );
+    const isConnected = Boolean(isSelected && activeConnection && activeConnection.getConnector() === activeConnector);
+    const isDisabled = Boolean(!isSelected && activeConnectorType && activeConnection);
     return { isSelected, isConnected, isDisabled, select };
 }