+++ samples/contractupdate/src/config.ts
@@ -1,0 +1,37 @@
+import { SignClientTypes } from '@walletconnect/types';
+import {
+    BrowserWalletConnector,
+    Network,
+    ephemeralConnectorType,
+    WalletConnectConnector,
+} from '@concordium/react-components';
+
+const TESTNET_GENESIS_BLOCK_HASH = '4221332d34e1694168c2a0c0b3fd0f273809612cb13d000d5c2e00e85f50f796';
+const MAINNET_GENESIS_BLOCK_HASH = '9dd9ca4d19e9393877d2c44b70f89acbfc0883c2243e5eeaecc0d1cd0503f478';
+
+export const TESTNET: Network = {
+    name: 'testnet',
+    genesisHash: TESTNET_GENESIS_BLOCK_HASH,
+    jsonRpcUrl: 'https://json-rpc.testnet.concordium.com',
+    ccdScanBaseUrl: 'https://testnet.ccdscan.io',
+};
+export const MAINNET: Network = {
+    name: 'mainnet',
+    genesisHash: MAINNET_GENESIS_BLOCK_HASH,
+    jsonRpcUrl: 'https://json-rpc.mainnet.concordium.software',
+    ccdScanBaseUrl: 'https://ccdscan.io',
+};
+
+const WALLET_CONNECT_PROJECT_ID = '76324905a70fe5c388bab46d3e0564dc';
+const WALLET_CONNECT_OPTS: SignClientTypes.Options = {
+    projectId: WALLET_CONNECT_PROJECT_ID,
+    metadata: {
+        name: 'Contract Update',
+        description: 'Example dApp for the performing an update on a contract.',
+        url: '#',
+        icons: ['https://walletconnect.com/walletconnect-logo.png'],
+    },
+};
+
+export const BROWSER_WALLET = ephemeralConnectorType(BrowserWalletConnector.create);
+export const WALLET_CONNECT = ephemeralConnectorType(WalletConnectConnector.create.bind(this, WALLET_CONNECT_OPTS));