+++ packages/react-components/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `WithWalletConnector`: Decouple component from concrete connector implementations by constructing instances from the application.
+    This also introduces the ability for applications to control the activation/deactivation lifecycle of the connectors.
+
 ## [0.1.0] - 2023-01-17
 
 ### Added